included all the security response headers in one filter class. Deleted ContentSecurityFilter and XFrameOptionsFilter
